(fix) Only the higher priority setting can change the RVC.
The current order for setting extensions:
1.  -march option (riscv_after_parse_args)
2.  md_begin (build arch hash table, and set the default text
    alignment according to rvc)
3.1 riscv_frag_align_code (handle alignment, according to rvc)
3.2 start parsing arch attribute before the first insn
3.3 .option rvc/norvc at any place

If the .align directive is set before .attribute directive, and
the -march option isn't set, then we may do the wrong alignment
behavior. I can not fix the problem so far...

Fortunately, our toolchain will generate -march option, and the
setting is same as the attribute directive.