gdb: riscv-contrib: fix NaN-boxing condition.
After different features(ABI and ISA) are used, call_info.flen is
currently from ABI feature (i.e. ELF), but the condition for NaN-boxing
should be based on the size of FPRs (i.e. ISA feature).

In this patch, no size of FPRs is checked. NaN-boxing is did only
when real FPRs are used to pass args and the type of passing arg
is float (i.e. size is 4).  This should work under Ext 'D' and Ext 'F'.