gdb: generic-contrib: fix form of label for abstract_origin
The form of label for abstract_origin should be a DW_FORM_ref4,
so that the actual position can be re-evaluated instead of a const
offset. In addition, the value for this form seems not to be a
negative value.

Modify this by using gdb.dwarf2/inlined_subroutine-inheritance.exp
as reference.