RISC-V Andes: fix R_RISCV_CALL as PLT issue (b27606)
the commit "Treat R_RISCV_CALL and R_RISCV_CALL_PLT as the same in check_relocs."
introduced an issue that the local jump/call targets are also referred
to as global calls which cause the target to be marked with the PLT flag.
when relocating these local jump/call relocations to a symbol with PLT attribute
might out of its scope.